+++ react-element/src/App.js
@@ -1,26 +1,105 @@
-import logo from './logo.svg';
 import './App.css';
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
+import React, { Component} from "react";
+
+export default class groceryArrList extends Component {
+  state = {
+    groceryArr: [],
+    inputValue: ""
+  };
+  price = {
+    //pull price from json file
+    priceArr: [],
+    priceVal: ""
+  };
+  investment = {
+    investArr: [],
+    investVal: ""
+  };
+
+  addGrocery = () => {
+    this.setState({ groceryArr: [this.state.inputValue, ...this.state.groceryArr] });
+    
+    // I want to insert separate paragraph tags (groceryArrs from this.state.groceryArr) into the list element here
+  };
+
+  handleKeyDown = (e) => {
+    if (e.keyCode === 13) {
+      if (this.state.inputValue === "") return;
+
+      this.addGrocery();
+      this.state.inputValue = "";
+    }
+  };
+
+  handleChange = (e) => {
+    this.setState({ inputValue: e.target.value });
+  };
+
+  render() {
+    const mystyle = {
+      color: "white",
+      backgroundColor: "DodgerBlue",
+      padding: "10px",
+      fontFamily: "Arial",
+      display: "flex",
+      flexDirection: "row"
+    };
+    const mainBox = {
+      display: "flex",
+      alignItems: "center",
+      flexDirection: "column"
+    };
+    const NPI = {
+      display: "flex",
+      alignItems: "column",
+      justifyContent: "Space-between"
+    };
+    const titleStyle = {
+      backgroundColor: "#B27092"
+
+    };
+    const listCategory = {
+      padding: "90px"
+    }
+    return (
+      <div style={mainBox}>
+        <h1 style={titleStyle}>
+          LETS START WHOPPING
+        </h1>
+        <div style={mystyle}>
+          <input
+            type="text"
+            className="groceryArr"
+            placeholder="Add a new item!"
+            onChange={this.handleChange}
+            onKeyDown={this.handleKeyDown}
+            value={this.state.inputValue}
+          />
+          <button onClick={this.groceryArr}>Add to list</button>
+        </div>
+        <div style={NPI}>
+          <div className="list" style={listCategory}>
+          <h3>
+            ITEM
+          </h3>
+            {this.state.groceryArr.map((groceryArr) => {
+              return <p>{groceryArr}</p>;
+            })}
+          </div>
+          <div style={listCategory}>
+            <h3>
+              PRICE
+            </h3>
+          </div>
+          <div style={listCategory}>
+            <h3>
+              INVESTMENT
+            </h3>
+          </div>
+        </div>
+
+      </div>
+    );
+  }
 }
-
-export default App;
-
